Fix import issue with Next.js 14, and tidy up
