+++ nextjs-end/package.json
@@ -9,10 +9,10 @@
 		"lint": "next lint"
 	},
 	"dependencies": {
-		"@google-cloud/firestore": "^6.7.0",
-		"firebase": "^10.3.1",
-		"firebase-admin": "^11.10.1",
-		"next": "13.4.10",
+		"@google-cloud/firestore": "^6.8.0",
+		"firebase": "^10.7.2",
+		"firebase-admin": "^11.11.1",
+		"next": "^14.1.0",
 		"protobufjs": "^7.2.5",
 		"react": "18.2.0",
 		"react-dom": "18.2.0",
