+++ nextjs-end/src/app/actions.js
@@ -1,7 +1,7 @@
 "use server";
 
 import { addReviewToRestaurant } from "@/src/lib/firebase/firestore.js";
-import { getAuthenticatedAppForUser } from "@/src/lib/firebase/firebase";
+import { getAuthenticatedAppForUser } from "@/src/lib/firebase/getAuthenticatedAppForUser.js";
 import { getFirestore } from "firebase/firestore";
 
 // This is a next.js server action, an alpha feature, so
