--- nextjs-end/src/lib/firebase/config.js
@@ -1,7 +1,0 @@
-export default {
-	projectId: "demo-codelab-nextjs",
-	appId: "demo-codelab-nextjs",
-	apiKey: "demo-codelab-nextjs",
-	storageBucket: "demo-codelab-nextjs.appspot.com",
-	authDomain: "demo-codelab-nextjs.firebaseapp.com",
-};
